Update README.md
Problem:
Predict the percentage of an student based on the no. of study hours.
What will be predicted score if a student studies for 9.25hrs/ day?
Simple Linear Regression :
In this task we will predict the percentage of marks that a student is expected to score based upon the number of hours they studied.
This is a simple linear regression task as it involves just two variables.